refactor: :recycle: extract MyApp in dedicated file
