+++ lib/main.dart
@@ -14,7 +14,7 @@ import 'package:breizh_blok_mobile/database/app_database.dart';
 import 'package:breizh_blok_mobile/image_boulder_cache.dart';
 import 'package:breizh_blok_mobile/location_provider.dart';
 import 'package:breizh_blok_mobile/models/order_param.dart';
-import 'package:breizh_blok_mobile/models/request_strategy.dart';
+import 'package:breizh_blok_mobile/my_app.dart';
 import 'package:breizh_blok_mobile/repositories/boulder_area_repository.dart';
 import 'package:breizh_blok_mobile/repositories/boulder_marker_repository.dart';
 import 'package:breizh_blok_mobile/repositories/boulder_repository.dart';
@@ -23,19 +23,11 @@ import 'package:breizh_blok_mobile/repositories/grade_repository.dart';
 import 'package:breizh_blok_mobile/repositories/municipality_repository.dart';
 import 'package:breizh_blok_mobile/share_content_service.dart';
 import 'package:breizh_blok_mobile/share_content_service_interface.dart';
-import 'package:breizh_blok_mobile/views/boulder_area_details_view.dart';
-import 'package:breizh_blok_mobile/views/boulder_details_view.dart';
-import 'package:breizh_blok_mobile/views/download_view.dart';
-import 'package:breizh_blok_mobile/views/home_view.dart';
-import 'package:breizh_blok_mobile/views/municipality_details_view.dart';
-import 'package:breizh_blok_mobile/views/profile_view.dart';
 import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
 import 'package:flutter/material.dart' hide HttpClientProvider;
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:go_router/go_router.dart';
 import 'package:location/location.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
@@ -175,123 +167,4 @@ Future<void> main({
       ),
     ),
   );
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({
-    required this.database,
-    super.key,
-  });
-
-  final AppDatabase database;
-
-  @override
-  Widget build(BuildContext context) {
-    final router = GoRouter(
-      routes: [
-        GoRoute(
-          path: '/',
-          name: 'boulder_list',
-          builder: (context, state) => RepositoryProvider(
-            create: (context) => RequestStrategy(),
-            child: HomeView(
-              database: database,
-            ),
-          ),
-          routes: [
-            GoRoute(
-              path: 'boulders/:id',
-              name: 'boulder_details',
-              builder: (context, state) {
-                return RepositoryProvider(
-                  create: (context) => RequestStrategy(),
-                  child: BoulderDetailsView(id: state.pathParameters['id']!),
-                );
-              },
-            ),
-            GoRoute(
-              path: 'boulder-areas/:id',
-              name: 'boulder_area_details',
-              builder: (context, state) {
-                return RepositoryProvider(
-                  create: (context) => RequestStrategy(),
-                  child:
-                      BoulderAreaDetailsView(id: state.pathParameters['id']!),
-                );
-              },
-            ),
-            GoRoute(
-              path: 'municipalities/:id',
-              name: 'municipality_details',
-              builder: (context, state) {
-                return RepositoryProvider(
-                  create: (context) => RequestStrategy(),
-                  child:
-                      MunicipalityDetailsView(id: state.pathParameters['id']!),
-                );
-              },
-            ),
-          ],
-        ),
-        GoRoute(
-          path: '/downloads/boulders/:id',
-          name: 'downloaded_boulder_details',
-          builder: (context, state) {
-            return RepositoryProvider<RequestStrategy>(
-              create: (context) => RequestStrategy(offlineFirst: true),
-              child: BoulderDetailsView(
-                id: state.pathParameters['id']!,
-                boulderAreaIri: state.uri.queryParameters['boulderAreaIri'],
-              ),
-            );
-          },
-        ),
-        GoRoute(
-          path: '/downloads/boulders-area/:id',
-          name: 'downloaded_boulder_area_details',
-          builder: (context, state) {
-            return RepositoryProvider<RequestStrategy>(
-              create: (context) => RequestStrategy(offlineFirst: true),
-              child: BoulderAreaDetailsView(
-                id: state.pathParameters['id']!,
-              ),
-            );
-          },
-        ),
-        GoRoute(
-          path: '/profile',
-          name: 'profile',
-          builder: (context, state) {
-            return RepositoryProvider(
-              create: (context) => RequestStrategy(),
-              child: const ProfileView(),
-            );
-          },
-        ),
-        GoRoute(
-          path: '/downloads',
-          name: 'downloads',
-          builder: (context, state) {
-            return RepositoryProvider(
-              create: (context) => RequestStrategy(),
-              child: DownloadView(
-                database: database,
-              ),
-            );
-          },
-        ),
-      ],
-      initialLocation: '/',
-    );
-
-    return MaterialApp.router(
-      routeInformationProvider: router.routeInformationProvider,
-      routeInformationParser: router.routeInformationParser,
-      routerDelegate: router.routerDelegate,
-      theme: ThemeData(useMaterial3: true, colorSchemeSeed: Colors.lightBlue),
-      debugShowCheckedModeBanner: false,
-      localizationsDelegates: AppLocalizations.localizationsDelegates,
-      supportedLocales: AppLocalizations.supportedLocales,
-    );
-  }
 }
